+++ admin/template/error_checking.php
@@ -21,6 +21,10 @@
       echo '<div class="error">'.i18n_r('WARNING').': <a href="health-check.php">'.i18n_r('SERVER_SETUP').' non-Apache</a></div>';
   }
 
+  if ($dataw->maintenance == '1') {
+	  echo '<div class="error">' . i18n_r('WARNING') . ': ' . i18n_r('MAINTENANCE_WARNING') . '</div>';
+  }
+
 	if(!isset($update)) $update = '';
 	$err = '';
 	$restored = '';