+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     * Added option to enable/disable code editor from settings page
     * Added option to enable/disable components
     * Added support for special templates for `403` and `404` errors handling pages
+    * Added **Maintenance Mode** to preventing visitors from using the site. This mode is enabled by default after installation of CMS
 
 1. [](#improved)
     * Updated theme function `get_navigation()` to add `aria-role="page"` to current page in menu